Change recipients to team alias
The mailing list seems to be working properly, so use it for recipients
of bot notifications. This reduces toil for the joiner/leaver process as
we can just curate the list behind that alias.

Note that this is still a list; if for some reason we needed to include
a recipient who wasn't in the mailing list, we could.