lambdas: make install-vars.json a "layer"
For some reason a few months ago the `input` from our
`cloudwatchevent_rule.targets[]` for both jobs stopped being included
in the event json, resulting in job failures due to `KeyError`s. With
this commit, we push install-vars.json as a "layer" instead, and tweak
the code to merge it into the event.